Refactored CfService into multiple services
