+++ src/main/java/org/tanzu/cfpulse/McpServerConfig.java
@@ -6,7 +6,7 @@ import org.springframework.ai.support.ToolCallbacks;
 import org.springframework.ai.tool.ToolCallback;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.tanzu.cfpulse.cf.CfService;
+import org.tanzu.cfpulse.cf.*;
 
 import java.util.List;
 
@@ -16,7 +16,15 @@ public class McpServerConfig {
     private static final Logger logger = LoggerFactory.getLogger(McpServerConfig.class);
 
     @Bean
-    public List<ToolCallback> registerTools(CfService cfService) {
-        return List.of(ToolCallbacks.from(cfService));
+    public List<ToolCallback> registerTools(
+            CfApplicationService cfApplicationService,
+            CfOrganizationService cfOrganizationService,
+            CfServiceInstanceService cfServiceInstanceService,
+            CfSpaceService cfSpaceService,
+            CfRouteService cfRouteService,
+            CfNetworkPolicyService cfNetworkPolicyService) {
+
+        return List.of(ToolCallbacks.from(cfApplicationService, cfOrganizationService, cfServiceInstanceService,
+                cfSpaceService, cfRouteService, cfNetworkPolicyService));
     }
 }
